requirements.txt
Flask
flask-cors
pandas
openai
mysql-connector-python
openpyxl
